PYIC-3609 move all processing to browser and use common CDN
